Add files via upload
One of the classic machine learning problems is the Titanic dataset. It offers details about the people who traveled on the Titanic, with the aim of estimating the likelihood of a passenger's survival based on a range of characteristics, including age, gender, class, and more. This project serves as a great introduction to feature engineering, model construction, and data cleansing.